Rename & New Prefixes
Renamed one of the 7 folder structure. And added a couple of new prefixes. Decided to create a -0a prefix for quarks. Used this prefix instead of the -q so that it can be ahead of quarks alphabetically. This might be temporary until I come up with a better prefixing system, but I decided to also add the -a after the 0 because the 0 might look similar to -o (organisms).

Added a -l structure for layouts